separate example file
